Revert "ScoreT is now a long"
This reverts commit 7445dbc9ddee152ee65575f65dfe786b65c90c14.

+++ lib/scorebot/Types.hpp
@@ -4,7 +4,7 @@
 #include <Arduino.h>
 
 using PlayerNumberT = int;
-using ScoreT = long;
+using ScoreT = int;
 using TurnNumberT = int;
 using TimestampT = decltype(millis());
